requirements.txt
fastapi
uvicorn[standard]
playwright
requests
python-dotenv
beautifulsoup4
PyMuPDF
lxml
